Updated KosmicTaskServer to 32/64 bit. Language plugins not 64 bit capable so these currently will not load. This means that the script type cannot be validated. So tasks cannot be validated and do not load. App is stable though.

+++ KosmicTask/KosmicTask/source/mother framework/MGSServerScriptRequest.m
@@ -708,7 +708,7 @@ validate the script folder
 	[scriptManager setDictionary:scriptDict];
 	
 	// scan script array
-	int i, scriptCount;
+	NSInteger i, scriptCount;
 	scriptCount  = [scriptManager count];
 	for (i = scriptCount - 1; i >= 0; i--) {
 		MGSScript *script = [scriptManager itemAtIndex:i];
